requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
pytorch-lightning==2.5.0.post0
datasets==3.2.0
transformers==4.48.1
huggingface-hub==0.27.1
python-dotenv~=1.1.0
pillow==11.1.0
tqdm==4.67.1
poetry==2.0.1
captum==0.7.0
tensorboardx==2.6.2.2
pip==24.3.1
jupyter==1.1.1
attrs==24.2.0
wheel==0.45.1
protobuf==5.28.3
filelock==3.16.1
setuptools==70.0.0
pywin32==308
opencv-python==4.10.0.84
matplotlib==3.9.2
numpy==2.0.2
scipy==1.14.1
scikit-learn~=1.6.0
torchmetrics~=1.6.1
seaborn==0.13.2
streamlit~=1.45.1
pathlib~=1.0.1